Use a dynamic method name for rule/decl iteration
In version 6 the `eachRule` and `eachDecl` were removed, after they were
deprecated in version 5, in favor of `walkRules` and `walkDecls`. No
further change is needed, the implementation of the methods stayed the
same. To keep supporting older versions we check for existence of the
new version and fall back to the old version.